GCP Projects/day_08/ex_08/requirements.txt
google-cloud-bigquery==1.28.0
google-cloud-datastore==1.15.3
fpdf==1.7.2